services: ping: rename exceptions module to avoid import clash
