+++ core/services/ping/ping_exceptions.py
